+++ src/VaultSharp/V1/AuthMethods/Cert/ICertAuthMethod.cs
@@ -1,6 +1,6 @@
 ﻿namespace VaultSharp.V1.AuthMethods.Cert
 {
-    public interface ICertAuthBackend
+    public interface ICertAuthMethod
     {
     }
 }